Refactor audio to separate module
